Move video/notification tests to the new folder structure
